chocolate-curry-32 Improve / fix package item lookup - Refactor to TierCortex from tierStateUtils - Refactor TierContext to include TierCortex - Consolidate ClientTierListManager into TierListManager - Refactor relevant pages to use TierCortex
